New BoundaryConditions submodule feels so good.
