+++ src/BoundaryConditions/pressure_boundary_conditions.jl
@@ -1,0 +1,15 @@
+#####
+##### Boundary conditions on pressure are derived from boundary conditions
+##### on the north-south horizontal velocity, v.
+#####
+
+# Pressure boundary conditions are either zero flux (Neumann) or Periodic.
+PressureBC(::BC)  = BoundaryCondition(Flux, nothing)
+PressureBC(::PBC) = PeriodicBC()
+
+function PressureBoundaryConditions(vbcs)
+    x = CoordinateBoundaryConditions(PressureBC(vbcs.x.left), PressureBC(vbcs.x.right))
+    y = CoordinateBoundaryConditions(PressureBC(vbcs.y.left), PressureBC(vbcs.y.right))
+    z = CoordinateBoundaryConditions(PressureBC(vbcs.z.left), PressureBC(vbcs.z.right))
+    return (x=x, y=y, z=z)
+end
